pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "growatt-bridge"
version = "0.1.0"
description = "FastAPI bridge service wrapping Growatt OpenAPI V1 with safety layer for AI agent control"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "growattServer>=1.9.0",
    "requests>=2.28.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "httpx>=0.27",
    "pytest-asyncio>=0.24",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
